+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 1.1.2 (Unreleased)
 
+* __[fix]__ Fix a warning due to Rails 5 `ActionController::Parameters` not being a Hash (@rstankov)
 * __[fix]__ Ensure `sort_by` prefixes with table_name. (@andreasklinger)
 
 ## Version 1.1.1